Apply security for indexing rest services

+++ src/main/java/uk/ac/ebi/biostudies/controller/Index.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 import uk.ac.ebi.biostudies.api.util.Constants;
+import uk.ac.ebi.biostudies.auth.UserSecurity;
 import uk.ac.ebi.biostudies.service.IndexService;
 
 import static uk.ac.ebi.biostudies.api.util.Constants.JSON_UNICODE_MEDIA_TYPE;
@@ -25,6 +26,9 @@ public class Index {
     @Autowired
     IndexService indexService;
 
+    @Autowired
+    UserSecurity userSecurity;
+
     /**
      * Method handling HTTP GET requests. The returned object will be sent
      * to the client as "text/plain" media type.
@@ -33,22 +37,28 @@ public class Index {
      */
     @RequestMapping(value = "/index/reload/{filename}", produces = JSON_UNICODE_MEDIA_TYPE, method = RequestMethod.GET)
     public ResponseEntity<String> indexAll(String filename) throws Exception {
+        if(!userSecurity.isManager())
+            return new ResponseEntity<String>("{\"message\":\"forbidden\"}", HttpStatus.FORBIDDEN);
         if(filename == null || filename.isEmpty() || filename.equalsIgnoreCase(Constants.STUDIES_JSON_FILE) || filename.equalsIgnoreCase("default")) {
             indexService.clearIndex(false);
             filename = "";
         }
         indexService.updateFromJSONFile(filename);
         indexService.indexAll(filename);
-        return new ResponseEntity<String>("{\"message\":\"Indexing started\"}", HttpStatus.OK);
+        return new ResponseEntity<String>("{}", HttpStatus.OK);
     }
 
     @RequestMapping(value = "/index/clear", produces = JSON_UNICODE_MEDIA_TYPE, method = RequestMethod.GET)
     public ResponseEntity<String> clearIndex() throws Exception {
+        if(!userSecurity.isManager())
+            return new ResponseEntity<String>("{\"message\":\"forbidden\"}", HttpStatus.FORBIDDEN);
         indexService.clearIndex(true);
         return new ResponseEntity<String>("{\"message\":\"Index empty\"}", HttpStatus.OK);
     }
     @RequestMapping(value = "/index/delete/{accession}", produces = JSON_UNICODE_MEDIA_TYPE, method = RequestMethod.GET)
     public ResponseEntity<String> deleteDoc(@PathVariable(required=false) String accession) throws Exception {
+        if(!userSecurity.isManager())
+            return new ResponseEntity<String>("{\"message\":\"forbidden\"}", HttpStatus.FORBIDDEN);
         indexService.deleteDoc(accession);
         return new ResponseEntity<String>("{\"message\":\"Index empty\"}", HttpStatus.OK);
     }
